splitted settings.ts into 2 files
